+++ src/directives/emoji-picker-caret.directive.ts
@@ -1,22 +1,27 @@
-import { Directive, Output, EventEmitter, ElementRef } from '@angular/core';
+import { Directive, Output, EventEmitter, ElementRef, ChangeDetectorRef, ApplicationRef, NgZone } from '@angular/core';
 import { Subject } from "rxjs/Subject";
 import 'rxjs/add/operator/takeUntil';
 import 'rxjs/add/operator/distinctUntilChanged';
 
-@Directive({ 
+import { CaretEvent } from "../../src";
+
+@Directive({
   selector: '[emojiPickerCaretEmitter]',
   host: {
     '(keyup)': 'updateCaretPosition()',
     '(mouseup)': 'updateCaretPosition()',
+    '(selectstart)': 'updateCaretPosition()',
     '(focus)': 'updateCaretPosition()',
-    '(DOMSubtreeModified)': 'updateCaretPosition($event)'
+    '(DOMSubtreeModified)': 'updateCaretDueMutation($event)'
   }
 })
 export class EmojiPickerCaretDirective {
-  @Output('emojiPickerCaretEmitter') caretEmitter = new EventEmitter();
+  @Output('emojiPickerCaretEmitter') caretEmitter = new EventEmitter<CaretEvent>();
 
-  private _position = new Subject<{ caretOffset, caretRange }>();
-  private _destroyed = new Subject<boolean>();
+  private _caretEvent$ = new Subject<CaretEvent>();
+  private _destroyed$ = new Subject<boolean>();
+
+  private _lastCaretEvent: CaretEvent = CaretEvent.null;
 
   private _win;
   private _doc;
@@ -37,87 +42,49 @@ export class EmojiPickerCaretDirective {
     return this._win
   }
 
-  constructor(private _el: ElementRef) { }
-
-  ngOnInit() {
-    if (!this._el.nativeElement.getAttribute('contenteditable') && this._el.nativeElement.tagName !== 'INPUT') {
-      throw new Error('(emojiPickerPositionEmitter) should only work on contenteditable enabled or input elements');
-    }
-
-    this._position
-      .takeUntil(this._destroyed)
+  constructor(
+    private _el: ElementRef
+  ) {
+    this._caretEvent$
+      .takeUntil(this._destroyed$)
       .distinctUntilChanged((event1, event2) => {
-        if (
-          /** if range suddenly exists or disappears */
-          !event1.caretRange && event2.caretRange || 
-          event1.caretRange && !event2.caretRange ||
-          /** if caret offset has changed */
-          event1.caretOffset !== event2.caretOffset ||
-          /** if caret range has changed in these properties */
-          !this.compareRangeObject(event1.caretRange, event2.caretRange)
-        ) {
-          return false;
-        }
-
-        return true;
+        return CaretEvent.compare(event1, event2);
       })
-      .subscribe(event => this.caretEmitter.emit(event))
+      .subscribe((event: CaretEvent) => {
+        this.caretEmitter.emit(event);
+        this._lastCaretEvent = event.clone()
+      })
     ;
   }
 
-  compareRangeObject(r1, r2) {
-    for (let k in r1) {
-      if (r1[k] !== r2[k]) {
-        return false
-      }
+  ngOnInit() {
+    if (!this._el.nativeElement.getAttribute('contenteditable') && this._el.nativeElement.tagName.toLowerCase() !== 'input') {
+      throw new Error('(emojiPickerPositionEmitter) should only work on contenteditable enabled or input elements');
     }
-
-    return true;
   }
 
   ngOnDestroy() {
-    this._destroyed.next(true);
+    this._destroyed$.next(true);
   }
 
   updateCaretPosition() {
-    const position = this.getCaretCharacterOffsetWithin(this.win, this.doc, this._el.nativeElement);
-    this._position.next(position);
+    const cEvent = CaretEvent.generateCaretEvent(this.win, this.doc, this._el.nativeElement);
+    this._caretEvent$.next(cEvent);
   }
 
-  getCaretCharacterOffsetWithin(win, doc, element) {
-    let caretOffset = 0, sel, caretRange;
+  updateCaretDueMutation() {
+    const cEvent = CaretEvent.generateCaretEvent(this.win, this.doc, this._el.nativeElement);
+    let textMovement = cEvent.textContent.length - this._lastCaretEvent.textContent.length;
+    cEvent.caretOffset = this._lastCaretEvent.caretOffset + textMovement;
 
-    if (typeof win.getSelection != "undefined") {
-      sel = win.getSelection();
-      if (sel.rangeCount > 0) {
-        const range = win.getSelection().getRangeAt(0);
-        const preCaretRange = range.cloneRange();
-        preCaretRange.selectNodeContents(element);
-        preCaretRange.setEnd(range.endContainer, range.endOffset);
-        caretOffset = preCaretRange.toString().length;
+    /** change detection after DOMSubtreeModified event is weird
+     * ChangeDetectorRef.detectChanges(), ChangeDetectorRef.markForCheck(), ApplicationRef.tick(), NgZone.run()
+     * all of those methods did not work as expected.
+     * As a temporary hack I am emitting an event after a short timeout, which is fine due to the _caretEvent$ smart stream
+     */
 
-        /** Keeping a reference of the range to emit */
-        caretRange = range.cloneRange();
-      }
-    } else if ((sel = doc.selection) && sel.type != "Control") {
-      const textRange = sel.createRange();
-      const preCaretTextRange = doc.body.createTextRange();
-      preCaretTextRange.moveToElementText(element);
-      preCaretTextRange.setEndPoint("EndToEnd", textRange);
-      caretOffset = preCaretTextRange.text.length;
-
-      /** Keeping a reference of the range to emit and making it compatible */
-      caretRange = textRange.duplicate();
-      caretRange.insertNode = (e) => {
-        const container = document.createElement("div");
-        container.appendChild(e);
-        caretRange.pasteHTML(container.innerHTML);
-      };
-    }
-
-    return {
-      caretOffset,
-      caretRange
-    };
+     setTimeout(() => {
+       this._caretEvent$.next(cEvent);
+     });
   }
 }