+++ src/index.ts
@@ -1,1 +1,2 @@
 export * from './emoji-picker.module';
+export * from './lib';