+++ src/lib/index.ts
@@ -1,0 +1,2 @@
+export * from './caret-event';
+export * from './emoji-event';